feat: abstract set field and create response serializer
