+++ src/django_project/genre_app/serializers.py
@@ -1,25 +1,6 @@
 from rest_framework import serializers
 
-from src.django_project.serializers import ListResponseSerializer
-
-
-class SetField(serializers.ListField):
-    """
-    Serializer for set field
-    """
-
-    def to_internal_value(self, data):  # type: ignore
-        """
-        Override to_internal_value to convert the list of uuids to a set
-        """
-        return set(super().to_internal_value(data))
-
-    def to_representation(self, data):
-        """
-        Override to_representation to convert the set of uuids to a list.
-        """
-
-        return list(super().to_representation(data))
+from src.django_project.serializers import ListResponseSerializer, SetField
 
 
 class GenreResponseSerializer(serializers.Serializer):
@@ -54,14 +35,6 @@ class CreateGenreRequestSerializer(serializers.Serializer):
     name = serializers.CharField(max_length=255, allow_blank=False)
     is_active = serializers.BooleanField(default=True)
     categories = SetField(child=serializers.UUIDField(), required=False)
-
-
-class CreateGenreResponseSerializer(serializers.Serializer):
-    """
-    Serializer for create genre response
-    """
-
-    id = serializers.UUIDField()
 
 
 class UpdateGenreRequestSerializer(serializers.Serializer):
